requirements.txt
streamlit>=1.29.0
anthropic>=0.40.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-pptx>=0.6.21
openpyxl>=3.1.0
pandas>=2.0.0
